% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_grade_learnr.R
\name{py_grade_learnr}
\alias{py_grade_learnr}
\title{R wrapper around `python_grade_learnr`}
\usage{
py_grade_learnr(...)
}
\arguments{
\item{...}{All learnr arguments required for checking}
}
\value{
An R list which contains several fields indicating the result of the
  check.
}
\description{
R wrapper around `python_grade_learnr`
}
